AI_Resume_Screening/AI_Resume_Screening/requirements.txt
langchain
langchain-core
langchain-groq
python-dotenv
PyPDF2